Refacto Kinds & render
