+++ src/game/pieces/mod.rs
@@ -1,1 +1,2 @@
 pub mod components;
+pub mod spawners;
